fix: on reconnect do not disconnect the new client
Currently when a client connects that is already there, the old socket is closed, and the code tries to reuse the existing client object be exchanging
its socket.

Reusing the same client object and just changing its socket does cause issues though with copies of the client in other threads. In the situations
that I could reproduce, it always disconnected both sockets, the old one and then the new one.

Instead I make a copy of the client object now, remove the old object and add the new object to the collection.