only start listening for clients once everything is initialized
Otherwise clients might connect to the server before everything is ready for them.

E.g. when restarting the server, the clients will immediately try to reconnect.

Clients might connect before the `PacketHandler` is initialized, which results in some packets not being processed by the server.

Same goes for the commands: Discord or the JsonApi might send in commands before all commands were added to the `CommandHandler`.

Without the `ClientJoined` action, clients might even be allowed connect if they are on the banlist.